backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-ratelimit==4.1.0
django-csp==3.8
django-extensions==3.2.3
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
celery==5.4.0
celery[redis]==5.4.0
django-celery-beat==2.7.0
redis==5.1.1
mysqlclient==2.2.4
Pillow==10.4.0
bleach==6.1.0
reportlab==4.2.2
requests==2.32.3
beautifulsoup4==4.12.3
fake-useragent==1.5.1
selenium==4.25.0
webdriver-manager==4.0.2
gunicorn==23.0.0
uvicorn[standard]==0.31.0
whitenoise==6.7.0
python-magic==0.4.27
python-dotenv==1.0.1
spacy==3.7.6
sentence-transformers==3.1.1
transformers==4.44.2
torch==2.4.1
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.1
